Fixed a few warnings.

+++ src/fsort.c
@@ -91,7 +91,7 @@ void fsort(void *base,
 #ifdef HANDLES_UNALIGNED_MEMORY_ACCESS
   switch(elmSize)
 #else
-  switch( (((unsigned long)base) % elmSize) ? 0 : elmSize )
+  switch( (((size_t)base) % elmSize) ? 0 : elmSize )
 #endif
   {
   case  1:  fsort_1(( B1_T *)base,(elms-1)+( B1_T *)base, cmpfunc); break;
